testutil: extract shared CaptureProcessor
